feat(bigquery): refactor test class of encryption configuration and change location in key
